manifests/bridges/bluetooth2kafka/src/requirements.txt
APScheduler 
bleak
paho-mqtt
kafka-python
pyyaml